build: activity > fragment 수정
