+++ feature/src/main/java/com/sopt/now/feature/mypage/MyPageFragment.kt
@@ -3,9 +3,7 @@ package com.sopt.now.feature.mypage
 import androidx.fragment.app.viewModels
 import com.sopt.now.core.base.BindingFragment
 import com.sopt.now.core.util.fragment.toast
-import com.sopt.now.core.util.intent.navigateTo
 import com.sopt.now.feature.R
-import com.sopt.now.feature.auth.LoginActivity
 import com.sopt.now.feature.databinding.FragmentMyPageBinding
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -26,7 +24,7 @@ class MyPageFragment : BindingFragment<FragmentMyPageBinding>(R.layout.fragment_
         binding.tvMainSignOut.setOnClickListener {
             viewModel.updateCheckLoginState(false)
             toast(getString(R.string.login_completed, getString(R.string.main_logout_under_bar)))
-            navigateTo<LoginActivity>(requireContext())
+            // TODO 로그인 프레그먼트로 이동
         }
     }
 
@@ -39,7 +37,7 @@ class MyPageFragment : BindingFragment<FragmentMyPageBinding>(R.layout.fragment_
                     getString(R.string.main_clear_user_under_bar)
                 )
             )
-            navigateTo<LoginActivity>(requireContext())
+            // TODO 로그인 프레그먼트로 이동
         }
     }
